Refactored where IdModel class is defined. Add basic ServiceCollection extension methods.
