+++ src/D20Tek.Country.Shared/D20Tek.Country.Shared.projitems
@@ -10,6 +10,5 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Models\CountryModel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\IdModel.cs" />
   </ItemGroup>
 </Project>
