+++ src/D20Tek.CountryApi/Common/IdModel.cs
@@ -1,7 +1,7 @@
 ﻿//---------------------------------------------------------------------------------------------------------------------
 // Copyright (c) d20Tek.  All rights reserved.
 //---------------------------------------------------------------------------------------------------------------------
-namespace D20Tek.Country.Shared.Models
+namespace D20Tek.CountryApi.Common
 {
     public abstract class IdEntity
     {
